Fix performance issue with dataLoader by moving registerNotification to AppProvider
